Working copy: correct type inference warning

+++ Blocstagram/Media.h
@@ -30,7 +30,7 @@ typedef NS_ENUM(NSInteger, MediaDownloadState) {
 @property (nonatomic, strong) NSArray *comments;
 @property (nonatomic, assign) MediaDownloadState downloadState;
 @property (nonatomic, assign) LikeState likeState;
-@property (nonatomic, assign) NSInteger numberOfLikes;
+@property (nonatomic, assign) int numberOfLikes;
 
 - (instancetype)initWithDictionary:(NSDictionary *)mediaDictionary;
